jh/requirements.txt
selenium==4.18.1
pandas==2.2.2
python-dotenv==1.0.1
cryptography==42.0.5
PyYAML==6.0.1
webdriver-manager==4.0.1